out_rdkafka2: add patch for new version of rdkafka
